* moving getTransport util.go -> connection.go * adding ability to handle custom Transporter in getTransport * adding test for the above scenario
